feat: throw ResourceAlreadyExistsException if dump fixture already exists

+++ src/Generators/AbstractTestsGenerator.php
@@ -6,6 +6,7 @@ use DateTime;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use RonasIT\Support\Events\SuccessCreateMessage;
+use RonasIT\Support\Exceptions\ResourceAlreadyExistsException;
 use RonasIT\Support\Exceptions\CircularRelationsFoundedException;
 
 abstract class AbstractTestsGenerator extends EntityGenerator
@@ -60,6 +61,10 @@ abstract class AbstractTestsGenerator extends EntityGenerator
         $this->createFixtureFolder();
 
         $dumpName = $this->getDumpName();
+
+        if(file_exists($this->getFixturesPath($dumpName))) {
+            throw new ResourceAlreadyExistsException($this->getFixturesPath($dumpName));
+        }
 
         file_put_contents($this->getFixturesPath($dumpName), $content);
